Initial code
This is the first upload of the code. I have worked on this for a few weeks on my computer but will now use GitHub for version control and experiments, exploring different ideas I have for this project. The following is a brief description of each .js file included and their purposes (note this might be subject to change): 

simulation.js: Main program that will create the variables and run the simulation.
vehicle.js: Contains the vehicle class with functions that determine its movement and other properties in the future
path.js: Contains the path class. This class is somewhat unnecessary for the moment. It simply draws the paths on the canvas. Future properties will be explore and the relevance of this separate file will be reevaluated. 

For more info on the project, see READ ME.